requirements.txt
alembic==0.9.10
aniso8601==2.0.0
click==6.7
Flask==0.12.2
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
MySQL-python==1.2.5
Pillow==5.2.0
python-dateutil==2.7.3
python-editor==1.0.3
python-magic==0.4.15
pytz==2017.3
PyYAML==3.13
six==1.11.0
SQLAlchemy==1.2.0
Werkzeug==0.14.1
